Use valve in endpoint
